refactoring : option btn component
Relate to : #156
